+++ src/hx/libs/asys/libuv/BaseData.cpp
@@ -1,0 +1,4 @@
+#include <hxcpp.h>
+#include "BaseData.h"
+
+hx::asys::libuv::BaseData::~BaseData() {}
