+++ src/hx/libs/asys/libuv/BaseData.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace hx::asys::libuv
+{
+    class BaseData
+    {
+    public:
+        virtual ~BaseData() = 0;
+    };
+}
