+++ src/hx/libs/asys/libuv/Build.xml
@@ -128,7 +128,9 @@
         <compilerflag value='-I${LIBUV_DIR}/include'/>
         <compilerflag value='-I${LIBUV_DIR}/src'/>
 
-        <file name='${this_dir}/LibuvAsys.cpp'/>
+        <file name='${this_dir}/LibuvAsysContext.cpp'/>
+        <file name='${this_dir}/Event.cpp'/>
+        <file name='${this_dir}/BaseData.cpp'/>
     </files>
 
     <target id='haxe'>
